+++ src/Parser/Evaluated/Rule/Natural/Divider.php
@@ -7,9 +7,9 @@ namespace Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural;
 
 use Serps\Core\Serp\IndexedResultSet;
 use Serps\SearchEngine\Google\Page\GoogleDom;
-use Serps\SearchEngine\Google\Parser\ParsingRuleInterace;
+use Serps\SearchEngine\Google\Parser\ParsingRuleInterface;
 
-class Divider implements \Serps\SearchEngine\Google\Parser\ParsingRuleInterace
+class Divider implements \Serps\SearchEngine\Google\Parser\ParsingRuleInterface
 {
 
     public function match(GoogleDom $dom, \DOMElement $node)