+++ src/Parser/Evaluated/Rule/Natural/VideoGroup.php
@@ -9,12 +9,12 @@ use Serps\Core\Serp\BaseResult;
 use Serps\Core\Serp\IndexedResultSet;
 use Serps\SearchEngine\Google\NaturalResultType;
 use Serps\SearchEngine\Google\Page\GoogleDom;
-use Serps\SearchEngine\Google\Parser\ParsingRuleInterace;
+use Serps\SearchEngine\Google\Parser\ParsingRuleInterface;
 
 /**
  * This rule extracts video groups as present on mobile results
  */
-class VideoGroup implements ParsingRuleInterace
+class VideoGroup implements ParsingRuleInterface
 {
 
     public function match(GoogleDom $dom, \DOMElement $node)